Refactor toward removing duplication on waiting on API responses.
